+++ OutOfOffice.DAL/Entity/Employees/GeneralEmployee.cs
@@ -11,7 +11,7 @@ public class GeneralEmployee : BaseEmployeeEntity
     public Position Position { get; set; }= null!;
     
     public bool Status { get; set; }
-    public decimal OutOfOfficeBalance { get; set; }
+    public int OutOfOfficeBalance { get; set; }
 
     public int? HrMangerId { get; set; }
     public HrManager HrManager { get; set; } = null!;
